+++ ChangeLog.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (Documentation) Many minor fixes like XML doc linter errors/warnings
 - (Export) Removed redundant texture entries in glTF schema
 - (Export) Properly closing buffer file stream
+- (Export) Conflict of textures with identical names
 
 ## [4.5.0] - 2022-01-24
 ### Added